Add ARMv7-M (PMSAv7) IPC test with proper app_package infrastructure
- Created complete IPC test structure in pw_kernel/target/armv7m_minimal/ipc/user/
  - BUILD.bazel: system_image with initiator/handler apps
  - target.rs: Kernel entry point for ARMv7-M IPC test
  - system.json5: Array-format config with IPC channel objects

- Updated target.ld.jinja with complete linker script sections
  - Added .code, .ARM, .FLASH.unused_space sections
  - Copied from MPS2 template to fix linker errors

- Added entry.rs platform entry point

- Enhanced system_generator to support array-format system.json5
  - Apps and objects now accept array format: [{name: '...'}]
  - Maintains backward compatibility with map format

- Updated armv7m_minimal platform configuration
  - Increased RAM size to 32768 bytes
  - Added linker_script_template filegroup export

- ARMv7-M IPC test validated in QEMU:
  - No ELF overlap errors
  - Successful kernel boot and userspace initialization
  - Working IPC communication between processes
  - Equivalent functionality to ARMv8-M MPS2

+++ pw_kernel/arch/arm_cortex_m/protection_v7.rs
@@ -264,7 +264,7 @@ impl MemoryConfig {
                 .with_privdefena(true),
         );
 
-        pw_log::info!("Programming {} MPU regions (PMSAv7)", self.mpu_regions.len());
+        pw_log::info!("Programming {} MPU regions (PMSAv7)", self.mpu_regions.len() as u32);
         
         for (index, region) in self.mpu_regions.iter().enumerate() {
             region.write(&mut mpu, index);
